requirements.txt
chromadb
google-generativeai
langchain
langchain-google-genai
langchain-text-splitters
pandas
opencv-python
Pillow
openai-whisper
pydub
ffmpeg-python
python-dotenv
pypdf2
tabula-py
jpype1
numpy
requests
tqdm
fastapi
uvicorn
python-multipart